refactor: extract LaTeX preview logic to separate module
Move LaTeX preview and diff operations to src/tools/approval/latexPreview.ts:
- previewProposedLatex() - preview proposed LaTeX document
- runLatexdiff() - run latexdiff and display result
- createTempFile() - create temp files with configurable location
- cleanupLatexAuxFiles() - clean up LaTeX auxiliary files

This reduces toolEditApproval.ts from ~1015 to ~793 lines and improves
separation of concerns. The main file now focuses on approval logic
while LaTeX operations are encapsulated in the new module.

Introduces LatexPreviewEntry interface for type-safe entry handling.
